GenomicBreeding

Documentation for GenomicBreeding.

GenomicBreeding.Genomes

GenomicBreeding.check

GenomicBreeding.Genomes — Type

Genomes struct containing unique entries and loci where allele frequencies can have missing values

Fields

entries: names of the n entries or samples

loci: names of the p loci including the chromsome or scaffold name, position, all alleles, and current allele separated by tabs

allele_frequencies: n x p matrix of allele frequencies between 0 and 1 which can have missing values

mask: n x p matrix of boolean mask for selective analyses and slicing

Examples

julia> g = Genomes(["entry_1", "entry_2"], ["chr1_12345_A|T_A", "chr2_678910_C|D_D"], [0.5 0.25; 0.9 missing], [true true; true false])
Genomes(["entry_1", "entry_2"], ["chr1_12345_A|T_A", "chr2_678910_C|D_D"], Union{Missing, Real}[0.5 0.25; 0.9 missing], Bool[1 1; 1 0])

julia> fieldnames(Genomes)
(:entries, :loci, :allele_frequencies, :mask)

GenomicBreeding.check — Method

Check dimension compatibility of the fields of the Genomes struct

Examples

julia> g = Genomes(["entry_1", "entry_2"], ["chr1_12345_A|T_A", "chr2_678910_C|D_D"], [0.5 0.25; 0.9 missing], [true true; true false]);

julia> check(g)
true

julia> g.entries = ["beaking_change"];

julia> check(g)
false