GenomicBreeding

Documentation for GenomicBreeding.

GenomicBreeding.Genomes

GenomicBreeding.Phenomes

GenomicBreeding.Trials

GenomicBreeding.check

GenomicBreeding.check

GenomicBreeding.check

GenomicBreeding.simulategenomes

GenomicBreeding.Genomes — Type

Genomes struct containing unique entries and loci where allele frequencies can have missing values

Fields

entries: names of the n entries or samples

loci: names of the p loci including the chromsome or scaffold name, position, all alleles, and current allele separated by tabs

allele_frequencies: n x p matrix of allele frequencies between 0 and 1 which can have missing values

mask: n x p matrix of boolean mask for selective analyses and slicing

Examples

julia> genomes = Genomes(n=2, p=2)
Genomes([#undef, #undef], [#undef, #undef], Union{Missing, Real}[#undef #undef; #undef #undef], Bool[0 0; 0 0])

julia> fieldnames(Genomes)
(:entries, :loci, :allele_frequencies, :mask)

julia> genomes.entries = ["entry_1", "entry_2"];

julia> genomes.loci = ["chr1_12345_A|T_A", "chr2_678910_C|D_D"];

julia> genomes.allele_frequencies = [0.5 0.25; 0.9 missing];

julia> genomes.mask = [true true; true false];

julia> genomes
Genomes(["entry_1", "entry_2"], ["chr1_12345_A|T_A", "chr2_678910_C|D_D"], Union{Missing, Real}[0.5 0.25; 0.9 missing], Bool[1 1; 1 0])

GenomicBreeding.Phenomes — Type

Phenomes struct containing unique entries and traits where phenotype data can have missing values

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

traits: names of the t traits

phenotypes: n x t matrix of numeric (R) phenotype data which can have missing values

mask: n x t matrix of boolean mask for selective analyses and slicing

Examples

julia> phenomes = Phenomes(n=2, t=2)
Phenomes([#undef, #undef], [#undef, #undef], [#undef, #undef], Union{Missing, Real}[#undef #undef; #undef #undef], Bool[0 0; 0 0])

julia> fieldnames(Phenomes)
(:entries, :populations, :traits, :phenotypes, :mask)

julia> phenomes.entries = ["entry_1", "entry_2"];

julia> phenomes.populations = ["pop_A", "pop_B"];

julia> phenomes.traits = ["height", "yield"];

julia> phenomes.phenotypes = [200.0 2.5; 150.0 missing];

julia> phenomes.mask = [true true; true false];

julia> phenomes
Phenomes(["entry_1", "entry_2"], ["pop_A", "pop_B"], ["height", "yield"], Union{Missing, Real}[200.0 2.5; 150.0 missing], Bool[1 1; 1 0])

GenomicBreeding.Trials — Type

Trials struct containing phenotype data across years, seasons, harvest, sites, populations, replications, blocks, rows, and columns

Fields

entries: names of the n entries or samples

traits: names of the t traits

phenotypes: n x t matrix of numeric (R) phenotype data which can have missing values

mask: n x t matrix of boolean mask for selective analyses and slicing

Examples

julia> trials = Trials(n=1, t=2)
Trials([#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], Union{Missing, Real}[#undef #undef])

julia> fieldnames(Trials)
(:traits, :years, :seasons, :harvests, :sites, :populations, :replications, :rows, :cols, :entries, :phenotypes)

GenomicBreeding.check — Method

Check dimension compatibility of the fields of the Genomes struct

Examples

julia> g = Genomes(n=2,p=4);

julia> check(g)
true

julia> g.entries = ["beaking_change"];

julia> check(g)
false

GenomicBreeding.check — Method

Check dimension compatibility of the fields of the Phenomes struct

Examples

julia> y = Phenomes(n=2, t=2);

julia> check(y)
true

julia> y.populations = ["beaking_change"];

julia> check(y)
false

GenomicBreeding.check — Method

Check dimension compatibility of the fields of the Trials struct

Examples

```jldoctest; setup = :(using GenomicBreeding)

julia> trials = Trials(n=1, t=2);

julia> trials.entries = ["entry_1"];

julia> trials.check() true

julia> trials.entries = ["entering2entries", "insteadofjust_1"];

julia> check(trials) false ```

GenomicBreeding.simulategenomes — Method

Simulate genomes

Arguments

n: number of entries (default = 100)

p: number of loci (default = 10_000)

n_chroms: number of chromosomes (default = 7)

n_alleles: number of alleles per locus (default = 2)

maxpos: total length of the genome in base-pairs (bp) (default = 135000_000)

ldcorr50perckb: distance in bp at which linkage expressed as correlation between a pair of loci is at 50% (default = 100000)

seed: psuedo-random number generator seed for replicability (default = 42)

verbose: Show progress bar and plot the linkage heatmap into an svg file? (default = true)

Output

Genomes

Examples

julia> genomes = simulategenomes(n=100, p=10_000, n_alleles=3, verbose=false);

julia> length(genomes.entries)
100

julia> length(genomes.loci)
20000

julia> size(genomes.allele_frequencies)
(100, 20000)

julia> idx = StatsBase.sample(range(1, 20_000, step=2), 250, replace = false, ordered = true);

julia> correlations = StatsBase.cor(genomes.allele_frequencies[:, idx]);

julia> correlations[10,10] == 1.00
true

julia> correlations[10,10] > correlations[10,250]
true