GenomicBreeding

Documentation for GenomicBreeding.

GenomicBreeding.Genomes

GenomicBreeding.Phenomes

GenomicBreeding.Trials

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.dimensions

GenomicBreeding.simulategenomes

GenomicBreeding.simulategenomiceffects

GenomicBreeding.simulatetrials

GenomicBreeding.Genomes — Type

Genomes struct

Containes unique entries and loci_alleles where allele frequencies can have missing values

Constructor

Genomes(; n::Int = 1, p::Int = 2)

Fields

entries: names of the n entries or samples

loci_alleles: names of the p loci-alleles combinations (p = l loci x a-1 alleles) including the chromsome or scaffold name, position, all alleles, and current allele separated by tabs

allele_frequencies: n x p matrix of allele frequencies between 0 and 1 which can have missing values

mask: n x p matrix of boolean mask for selective analyses and slicing

Examples

julia> genomes = Genomes(n=2, p=2)
Genomes([#undef, #undef], [#undef, #undef], Union{Missing, Real}[#undef #undef; #undef #undef], Bool[0 0; 0 0])

julia> fieldnames(Genomes)
(:entries, :loci_alleles, :allele_frequencies, :mask)

julia> genomes.entries = ["entry_1", "entry_2"];

julia> genomes.loci_alleles = ["chr1_12345_A|T_A", "chr2_678910_C|D_D"];

julia> genomes.allele_frequencies = [0.5 0.25; 0.9 missing];

julia> genomes.mask = [true true; true false];

julia> genomes
Genomes(["entry_1", "entry_2"], ["chr1_12345_A|T_A", "chr2_678910_C|D_D"], Union{Missing, Real}[0.5 0.25; 0.9 missing], Bool[1 1; 1 0])

GenomicBreeding.Phenomes — Type

Phenomes struct

Constains unique entries and traits where phenotype data can have missing values

Constructor

Phenomes(; n::Int = 1, t::Int = 2)

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

traits: names of the t traits

phenotypes: n x t matrix of numeric (R) phenotype data which can have missing values

mask: n x t matrix of boolean mask for selective analyses and slicing

Examples

julia> phenomes = Phenomes(n=2, t=2)
Phenomes([#undef, #undef], [#undef, #undef], [#undef, #undef], Union{Missing, Real}[#undef #undef; #undef #undef], Bool[0 0; 0 0])

julia> fieldnames(Phenomes)
(:entries, :populations, :traits, :phenotypes, :mask)

julia> phenomes.entries = ["entry_1", "entry_2"];

julia> phenomes.populations = ["pop_A", "pop_B"];

julia> phenomes.traits = ["height", "yield"];

julia> phenomes.phenotypes = [200.0 2.5; 150.0 missing];

julia> phenomes.mask = [true true; true false];

julia> phenomes
Phenomes(["entry_1", "entry_2"], ["pop_A", "pop_B"], ["height", "yield"], Union{Missing, Real}[200.0 2.5; 150.0 missing], Bool[1 1; 1 0])

GenomicBreeding.Trials — Type

Trials struct

Contains phenotype data across years, seasons, harvest, sites, populations, replications, blocks, rows, and columns

Constructor

Trials(; n::Int = 2, p::Int = 2)

Fields

phenotypes: n x t matrix of numeric phenotype data which can have missing values

traits: names of the traits t traits

years: names of the years corresponding to each row in the phenotype matrix

seasons: names of the seasons corresponding to each row in the phenotype matrix

harvests: names of the harvests corresponding to each row in the phenotype matrix

sites: names of the sites corresponding to each row in the phenotype matrix

replications: names of the replications corresponding to each row in the phenotype matrix

blocks: names of the blocks corresponding to each row in the phenotype matrix

rows: names of the rows corresponding to each row in the phenotype matrix

cols: names of the cols corresponding to each row in the phenotype matrix

entries: names of the entries corresponding to each row in the phenotype matrix

populations: names of the populations corresponding to each row in the phenotype matrix

Examples

julia> trials = Trials(n=1, t=2)
Trials(Union{Missing, Real}[#undef #undef], [#undef, #undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef])

julia> fieldnames(Trials)
(:phenotypes, :traits, :years, :seasons, :harvests, :sites, :replications, :blocks, :rows, :cols, :entries, :populations)

GenomicBreeding.checkdims — Method

checkdims(g::Genomes)::Bool

Check dimension compatibility of the fields of the Genomes struct

Examples

julia> g = Genomes(n=2,p=4);

julia> checkdims(g)
true

julia> g.entries = ["beaking_change"];

julia> checkdims(g)
false

GenomicBreeding.checkdims — Method

checkdims(y::Phenomes)::Bool

Check dimension compatibility of the fields of the Phenomes struct

Examples

julia> y = Phenomes(n=2, t=2);

julia> checkdims(y)
true

julia> y.populations = ["beaking_change"];

julia> checkdims(y)
false

GenomicBreeding.checkdims — Method

checkdims(trials::Trials)::Bool

Check dimension compatibility of the fields of the Trials struct

Examples

julia> trials = Trials(n=1, t=2);

julia> trials.entries = ["entry_1"];

julia> checkdims(trials)
true

julia> trials.entries = ["entering_2_entries", "instead_of_just_1"];

julia> checkdims(trials)
false

GenomicBreeding.dimensions — Method

dimensions(g::Genomes)::Tuple{Int, Int, Int}

Count the number of entries, loci, and maximum number of alleles per locus in the Genomes struct

Examples

julia> g = simulategenomes(n=100, l=1_000, n_alleles=4);

julia> dimensions(g)
(100, 3000, 1000, 4)

GenomicBreeding.simulategenomes — Method

Simulate genomes

Arguments

n: number of entries (default = 100)

l: number of loci (default = 10_000)

n_chroms: number of chromosomes (default = 7)

n_alleles: number of alleles per locus (default = 2)

max_pos: total length of the genome in base-pairs (bp) (default = 135000000)

ld_corr_50perc_kb: distance in bp at which linkage expressed as correlation between a pair of loci is at 50% (default = 100_000)

seed: psuedo-random number generator seed for replicability (default = 42)

verbose: Show progress bar and plot the linkage heatmap into an svg file? (default = false)

Output

Genomes

Examples

julia> genomes = simulategenomes(n=100, l=10_000, n_alleles=3);

julia> length(genomes.entries)
100

julia> length(genomes.loci_alleles)
20000

julia> size(genomes.allele_frequencies)
(100, 20000)

julia> rng::TaskLocalRNG = Random.seed!(123);

julia> idx = StatsBase.sample(rng, range(1, 20_000, step=2), 250, replace = false, ordered = true);

julia> correlations = StatsBase.cor(genomes.allele_frequencies[:, idx]);

julia> correlations[10,10] == 1.00
true

julia> correlations[10,10] > correlations[10,250]
true

GenomicBreeding.simulategenomiceffects — Method

Simulate genomic effects

Simulate additive, dominance, and epistatic effects

Arguments

genomes: Genome struct includes the n entries x p loci-alleles combinations (p = l loci x a-1 alleles)

f_additive: proportion of the l loci with non-zero additive effects on the phenotype

f_dominance: proportion of the l*f_additive additive effects loci with additional dominance effects

f_epistasis: proportion of the l*f_additive additive effects loci with additional dominance effects

Outputs

Examples

```jldoctest; setup = :(using GenomicBreeding)

```

GenomicBreeding.simulatetrials — Method

Simulate trials

Arguments

Output

Examples