GenomicBreeding

Documentation for GenomicBreeding.

GenomicBreeding.GBInput

GenomicBreeding.assess

GenomicBreeding.extracteffects

GenomicBreeding.load

GenomicBreeding.submitslurmarrayjobs

GenomicBreeding.GBInput — Type

mutable struct GBInput
    fname_geno::String
    fname_pheno::String
    bulk_cv::Bool
    populations::Union{Nothing, String, Vector{String}}
    traits::Union{Nothing, String, Vector{String}}
    models::Any
    n_folds::Int64
    n_replications::Int64
    maf::Float64
    mtv::Float64
    n_iter::Int64,
    n_burnin::Int64,
    fname_out_prefix::String,
    verbose::Bool
end

Input struct (belongs to GBCore.AbstractGB type)

fname_geno: genotype file (see file format guide: TODO: {URL HERE})

fname_pheno: phenotype file (see file format guide: TODO: {URL HERE})

bulk_cv: perform cross-validation across all populations, i.e. disregard population grouping (Default = false)

populations: include only these populations (Default = nothing which means include all populations)

traits: include only these traits (Default = nothing which means include all traits)

models: include these model functions (Default = [ridge, bayesa]; see models list: TODO: {URL HERE})

n_folds: number of k partitions for k-fold cross-validation (Default = 5)

n_replications: number of replications for repeated k-fold cross-validation (Default = 5)

keep_all: keep all entries upon merging genomes and phenomes potentially resulting in sparsities in both structs? (Default = false)

maf: minimum allele frequency (Default = 0.05)

mtv: minimum trait variance (Default = 1e-7)

n_iter: number of Bayesian model fitting MCMC/HMC iteration (Default = 1_500)

fname_out_prefix: prefix of the output files which may include directory names (Default = "" which translates to ./GBOuput/output-<yyyymmddHHMMSS>-<3_digit_random_number>-)

n_burnin: number of initial Bayesian model fitting MCMC/HMC iterations to be excluded from the posterior distribution (Default = 500)

verbose: show messages (Default = true)

GenomicBreeding.assess — Method

assess(input::GBInput)::Tuple{Vector{String},Vector{String}}

Assess genomic prediction accuracy via replicated k-fold cross-validation. Outputs are saved as JLD2 (each containing a CV struct) and TXT (each line is a note on failed runs) files.

Example

julia> genomes = GBCore.simulategenomes(n=300, l=1_000, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> proportion_of_variance = fill(0.0, 9, 3); proportion_of_variance[1, :] .= 1.00; # 100% variance on the additive genetic effects

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, proportion_of_variance=proportion_of_variance, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;

julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, populations=["pop_1", "pop_3"], traits=["trait_1"], n_replications=2, n_folds=3, verbose=false);

julia> fnames_cvs, fnames_notes = assess(input);

julia> length(fnames_cvs) == 32, length(fnames_notes) == 0
(true, true)

GenomicBreeding.extracteffects — Method

extracteffects(input::GBInput)::Vector{String}

Extract allele effects by fitting the models without cross-validation

Example

julia> genomes = GBCore.simulategenomes(n=300, l=1_000, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> proportion_of_variance = fill(0.0, 9, 3); proportion_of_variance[1, :] .= 1.00; # 100% variance on the additive genetic effects

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, proportion_of_variance=proportion_of_variance, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;

julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, populations=["pop_1", "pop_3"], traits=["trait_1"], n_replications=2, n_folds=3, verbose=false);

julia> fname_effects = extracteffects(input);

julia> length(fname_effects) == 6
true

GenomicBreeding.load — Method

load(input::GBInput)::Tuple{Genomes, Phenomes}

Load, merge and filter genotype and phenotype data

Example

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;
    
julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, populations=["pop_1", "pop_3"], traits=["trait_1"], verbose=false);

julia> genomes, phenomes = load(input);

julia> length(unique(genomes.populations)) == length(unique(phenomes.populations)) == 2
true

julia> length(phenomes.traits) == 1
true

julia> rm.([fname_geno, fname_pheno]);

GenomicBreeding.submitslurmarrayjobs — Method

submitslurmarrayjobs(
    fname_geno::String,
    fname_pheno::String;
    models::Any = [ridge, bayesa],
    n_folds::Int64 = 5,
    n_replications::Int64 = 5,
    maf::Float64 = 0.05,
    mtv::Float64 = 1e-7,
    n_iter::Int64 = 1_500,
    n_burnin::Int64 = 500,
    fname_out_prefix::String = "",
    verbose::Bool = true,
)::Vector{GBInput}

TODO: submit Slurm array job

Example

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;
    
julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> inputs = submitslurmarrayjobs(fname_geno=fname_geno, fname_pheno=fname_pheno, verbose=false);

julia> length(inputs) == 24
true

julia> rm.([fname_geno, fname_pheno]);