GenomicBreeding

Documentation for GenomicBreeding.

GenomicBreeding.Genomes

GenomicBreeding.Phenomes

GenomicBreeding.Trials

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.dimensions

GenomicBreeding.simulateeffects

GenomicBreeding.simulategenomes

GenomicBreeding.simulategenomiceffects

GenomicBreeding.simulatetrials

GenomicBreeding.Genomes — Type

Genomes struct

Containes unique entries and loci_alleles where allele frequencies can have missing values

Constructor

Genomes(; n::Int = 1, p::Int = 2)

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

loci_alleles: names of the p loci-alleles combinations (p = l loci x a-1 alleles) including the chromsome or scaffold name, position, all alleles, and current allele separated by tabs

allele_frequencies: n x p matrix of allele frequencies between 0 and 1 which can have missing values

mask: n x p matrix of boolean mask for selective analyses and slicing

Examples

julia> genomes = Genomes(n=2, p=2)
Genomes([#undef, #undef], [#undef, #undef], [#undef, #undef], Union{Missing, Real}[#undef #undef; #undef #undef], Bool[0 0; 0 0])

julia> fieldnames(Genomes)
(:entries, :populations, :loci_alleles, :allele_frequencies, :mask)

julia> genomes.entries = ["entry_1", "entry_2"];

julia> genomes.populations = ["pop_1", "pop_1"];

julia> genomes.loci_alleles = ["chr1_12345_A|T_A", "chr2_678910_C|D_D"];

julia> genomes.allele_frequencies = [0.5 0.25; 0.9 missing];

julia> genomes.mask = [true true; true false];

julia> genomes
Genomes(["entry_1", "entry_2"], ["pop_1", "pop_1"], ["chr1_12345_A|T_A", "chr2_678910_C|D_D"], Union{Missing, Real}[0.5 0.25; 0.9 missing], Bool[1 1; 1 0])

GenomicBreeding.Phenomes — Type

Phenomes struct

Constains unique entries and traits where phenotype data can have missing values

Constructor

Phenomes(; n::Int = 1, t::Int = 2)

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

traits: names of the t traits

phenotypes: n x t matrix of numeric (R) phenotype data which can have missing values

mask: n x t matrix of boolean mask for selective analyses and slicing

Examples

julia> phenomes = Phenomes(n=2, t=2)
Phenomes([#undef, #undef], [#undef, #undef], [#undef, #undef], Union{Missing, Real}[#undef #undef; #undef #undef], Bool[0 0; 0 0])

julia> fieldnames(Phenomes)
(:entries, :populations, :traits, :phenotypes, :mask)

julia> phenomes.entries = ["entry_1", "entry_2"];

julia> phenomes.populations = ["pop_A", "pop_B"];

julia> phenomes.traits = ["height", "yield"];

julia> phenomes.phenotypes = [200.0 2.5; 150.0 missing];

julia> phenomes.mask = [true true; true false];

julia> phenomes
Phenomes(["entry_1", "entry_2"], ["pop_A", "pop_B"], ["height", "yield"], Union{Missing, Real}[200.0 2.5; 150.0 missing], Bool[1 1; 1 0])

GenomicBreeding.Trials — Type

Trials struct

Contains phenotype data across years, seasons, harvest, sites, populations, replications, blocks, rows, and columns

Constructor

Trials(; n::Int = 2, p::Int = 2)

Fields

phenotypes: n x t matrix of numeric phenotype data which can have missing values

traits: names of the traits t traits

years: names of the years corresponding to each row in the phenotype matrix

seasons: names of the seasons corresponding to each row in the phenotype matrix

harvests: names of the harvests corresponding to each row in the phenotype matrix

sites: names of the sites corresponding to each row in the phenotype matrix

replications: names of the replications corresponding to each row in the phenotype matrix

blocks: names of the blocks corresponding to each row in the phenotype matrix

rows: names of the rows corresponding to each row in the phenotype matrix

cols: names of the cols corresponding to each row in the phenotype matrix

entries: names of the entries corresponding to each row in the phenotype matrix

populations: names of the populations corresponding to each row in the phenotype matrix

Examples

julia> trials = Trials(n=1, t=2)
Trials(Union{Missing, Real}[#undef #undef], [#undef, #undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef], [#undef])

julia> fieldnames(Trials)
(:phenotypes, :traits, :years, :seasons, :harvests, :sites, :replications, :blocks, :rows, :cols, :entries, :populations)

GenomicBreeding.checkdims — Method

checkdims(genomes::Genomes)::Bool

Check dimension compatibility of the fields of the Genomes struct

Examples

julia> genomes = Genomes(n=2,p=4);

julia> checkdims(genomes)
true

julia> genomes.entries = ["beaking_change"];

julia> checkdims(genomes)
false

GenomicBreeding.checkdims — Method

checkdims(y::Phenomes)::Bool

Check dimension compatibility of the fields of the Phenomes struct

Examples

julia> y = Phenomes(n=2, t=2);

julia> checkdims(y)
true

julia> y.populations = ["beaking_change"];

julia> checkdims(y)
false

GenomicBreeding.checkdims — Method

checkdims(trials::Trials)::Bool

Check dimension compatibility of the fields of the Trials struct

Examples

julia> trials = Trials(n=1, t=2);

julia> trials.entries = ["entry_1"];

julia> checkdims(trials)
true

julia> trials.entries = ["entering_2_entries", "instead_of_just_1"];

julia> checkdims(trials)
false

GenomicBreeding.dimensions — Method

dimensions(genomes::Genomes)::Tuple{Int, Int, Int}

Count the number of entries, populations, loci, and maximum number of alleles per locus in the Genomes struct

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> dimensions(genomes)
(100, 100, 3000, 1000, 4)

GenomicBreeding.simulateeffects — Method

Simulate effects

Sample p x q effects from a multivariate normal distribution with μ~Exp(λ) and Σ=μμ'

Arguments

p: number of correlated effects to simulate (default = 2)

q: number times to simulate the correlated effects from the same distribution (default = 1)

λ: parameter of the exponential distritbution from which the means will be sampled from (default = 1.00)

seed: randomisation seed (default = 42)

Output

p x q matrix of correlated effects

Examples

julia> θ::Array{Real,2} = simulateeffects();

julia> sum(abs.(θ - [-0.0886501800782904; -0.596478483888422])) < 0.00001
true

GenomicBreeding.simulategenomes — Method

Simulate genomes

Arguments

n: number of entries (default = 100)

l: number of loci (default = 10_000)

n_chroms: number of chromosomes (default = 7)

n_alleles: number of alleles per locus (default = 2)

max_pos: total length of the genome in base-pairs (bp) (default = 135000000)

ld_corr_50perc_kb: distance in bp at which linkage expressed as correlation between a pair of loci is at 50% (default = 100_000)

seed: psuedo-random number generator seed for replicability (default = 42)

verbose: Show progress bar and plot the linkage heatmap into an svg file? (default = true)

Output

Genomes

Examples

julia> genomes = simulategenomes(n=100, l=10_000, n_alleles=3, verbose=false);

julia> length(genomes.entries)
100

julia> length(genomes.populations)
100

julia> length(genomes.loci_alleles)
20000

julia> size(genomes.allele_frequencies)
(100, 20000)

julia> rng::TaskLocalRNG = Random.seed!(123);

julia> idx = StatsBase.sample(rng, range(1, 20_000, step=2), 250, replace = false, ordered = true);

julia> correlations = StatsBase.cor(genomes.allele_frequencies[:, idx]);

julia> correlations[10,10] == 1.00
true

julia> correlations[10,10] > correlations[10,250]
true

GenomicBreeding.simulategenomiceffects — Method

Simulate genomic effects

Simulate additive, dominance, and epistatic effects

Arguments

genomes: Genome struct includes the n entries x p loci-alleles combinations (p = l loci x a-1 alleles)

f_additive: proportion of the l loci with non-zero additive effects on the phenotype

f_dominance: proportion of the l*f_additive additive effects loci with additional dominance effects

f_epistasis: proportion of the l*f_additive additive effects loci with additional epistasis effects

Outputs

n x 3 matrix of additive, dominance and epistasis effects per entry

p x 3 matrix of additive, dominance and epistasis effects per locus-allele combination

Examples

julia> genomes::Genomes = simulategenomes(n=100, l=2_000, n_alleles=3, verbose=false);

julia> G, B = simulategenomiceffects(genomes=genomes, f_additive=0.05, f_dominance=0.75, f_epistasis=0.25);

julia> size.([G, B])
2-element Vector{Tuple{Int64, Int64}}:
 (100, 3)
 (4000, 3)

julia> sum(B .!= 0.0, dims=1)
1×3 Matrix{Int64}:
 200  75  50

Details

The additive, dominance, and epistasis allele effects share a common exponential distribution (λ=1) from which the mean of the effects (μ) are sampled, and the covariance matrix is derived (Σ = μ * μ'; where if det(Σ)≈0 then we iteratively add 1.00 to the diagonals until it becomes invertible or 10 iterations finishes and throws an error). The non-additive or epistasis allele effects were simulated by multiplying the allele frequencies of all possible unique pairs of epistasis alleles and their effects.

GenomicBreeding.simulatetrials — Method

Simulate trials

Arguments

genomes: Genome struct includes the n entries x p loci-alleles combinations (p = l loci x a-1 alleles)

n_traits: Number of traits (default = 3)

n_years: Number of years (default = 2)

n_seasons: Number of seasons (default = 4)

n_harvests: Number of harvests (default = 2)

n_sites: Number of sites (default = 4)

n_replications: Number of replications (default = 2)

n_blocks: Number of blocks across the entire field layout (default = missing)

n_rows: Number of rows across the entire field layout (default = missing)

n_cols: Number of columns across the entire field layout (default = missing)

proportion_of_variance: 12 x n_traits numeric matrix of scaled/non-scaled proportion of variances allocated to genetic and environmental effects (default = missing; values will be sampled from a uniform distribution followed by a biased sample on the first row, i.e. additive effects row). The rows correspond to the variance allocated to:

additive genetic effects

dominance genetic effects

epistasis genetic effects

years effects

seasons effects

harvests effects

sites effects

replications effects

blocks effects

rows effects

cols effects

complex interaction effects

seed: Randomisation seed (default = 42)

sparsity: Proportion of missing data (default = 0.0)

verbose: Show trials simulation progress bar? (default = true)

Output

Trials

Examples

julia> genomes::Genomes = simulategenomes(n=100, l=2_000, n_alleles=3, verbose=false);

julia> trials::Trials = simulatetrials(genomes=genomes, sparsity=0.25, verbose=false);

julia> size(trials.phenotypes)
(12800, 3)

julia> size(trials.traits)
(3,)

julia> unique(trials.entries) == genomes.entries
true

julia> unique(trials.populations) == unique(genomes.populations)
true

julia> abs(mean(ismissing.(trials.phenotypes)) - 0.25) < 0.00001
true