GenomicBreeding

Documentation for GenomicBreeding.

GenomicBreeding.Genomes

GenomicBreeding.Phenomes

GenomicBreeding.SimulatedEffects

GenomicBreeding.TEBV

GenomicBreeding.Trials

Base.:==

Base.:==

Base.:==

Base.:==

Base.:==

Base.hash

Base.hash

Base.hash

Base.hash

Base.hash

Base.sum

GenomicBreeding.analyse

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.checkdims

GenomicBreeding.countlevels

GenomicBreeding.dimensions

GenomicBreeding.loci

GenomicBreeding.loci_alleles

GenomicBreeding.plot

GenomicBreeding.readJLD2

GenomicBreeding.readdelimited

GenomicBreeding.simulateeffects

GenomicBreeding.simulategenomes

GenomicBreeding.simulategenomiceffects

GenomicBreeding.simulatetrials

GenomicBreeding.slice

GenomicBreeding.trialsmodelsfomulae!

GenomicBreeding.writeJLD2

GenomicBreeding.writedelimited

GenomicBreeding.writedelimited

GenomicBreeding.writedelimited

GenomicBreeding.@string2formula

GenomicBreeding.Genomes — Type

Genomes struct

Containes unique entries and loci_alleles where allele frequencies can have missing values

Constructor

Genomes(; n::Int64 = 1, p::Int64 = 2)

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

loci_alleles: names of the p loci-alleles combinations (p = l loci x a-1 alleles) including the

chromsome or scaffold name, position, all alleles, and current allele separated by tabs ("\t")

allele_frequencies: n x p matrix of allele frequencies between 0 and 1 which can have missing values

mask: n x p matrix of boolean mask for selective analyses and slicing

Examples

julia> genomes = Genomes(n=2, p=2)
Genomes(["", ""], ["", ""], ["", ""], Union{Missing, Float64}[missing missing; missing missing], Bool[0 0; 0 0])

julia> fieldnames(Genomes)
(:entries, :populations, :loci_alleles, :allele_frequencies, :mask)

julia> genomes.entries = ["entry_1", "entry_2"];

julia> genomes.populations = ["pop_1", "pop_1"];

julia> genomes.loci_alleles = ["chr1\t12345\tA|T\tA", "chr2\t678910\tC|D\tD"];

julia> genomes.allele_frequencies = [0.5 0.25; 0.9 missing];

julia> genomes.mask = [true true; true false];

julia> genomes
Genomes(["entry_1", "entry_2"], ["pop_1", "pop_1"], ["chr1\t12345\tA|T\tA", "chr2\t678910\tC|D\tD"], Union{Missing, Float64}[0.5 0.25; 0.9 missing], Bool[1 1; 1 0])

GenomicBreeding.Phenomes — Type

Phenomes struct

Constains unique entries and traits where phenotype data can have missing values

Constructor

Phenomes(; n::Int64 = 1, t::Int64 = 2)

Fields

entries: names of the n entries or samples

populations: name/s of the population/s each entries or samples belong to

traits: names of the t traits

phenotypes: n x t matrix of numeric (R) phenotype data which can have missing values

mask: n x t matrix of boolean mask for selective analyses and slicing

Examples

julia> phenomes = Phenomes(n=2, t=2)
Phenomes(["", ""], ["", ""], ["", ""], Union{Missing, Float64}[missing missing; missing missing], Bool[0 0; 0 0])

julia> fieldnames(Phenomes)
(:entries, :populations, :traits, :phenotypes, :mask)

julia> phenomes.entries = ["entry_1", "entry_2"];

julia> phenomes.populations = ["pop_A", "pop_B"];

julia> phenomes.traits = ["height", "yield"];

julia> phenomes.phenotypes = [200.0 2.5; 150.0 missing];

julia> phenomes.mask = [true true; true false];

julia> phenomes
Phenomes(["entry_1", "entry_2"], ["pop_A", "pop_B"], ["height", "yield"], Union{Missing, Float64}[200.0 2.5; 150.0 missing], Bool[1 1; 1 0])

GenomicBreeding.SimulatedEffects — Type

SimulatedEffects struct

Contains:

Identification, i.e. trait, year, season, harvest, site, and replication

Additive environmental effects:

year

season

site

Environmental interaction effects:

seasonsxyear

harvestsxseasonxyear

sitesxharvestxseasonxyear

Spatial effects including the field layout per year-season-harvest-site combination

field_layout

replicationsxsitexharvestxseasonxyear

blocksxsitexharvestxseasonxyear

rowsxsitexharvestxseasonxyear

colsxsitexharvestxseasonxyear

Genetic effects

additive_genetic

dominance_genetic

epistasis_genetic

GxE effects

additiveallelexsitexharvestxseasonx_year

dominanceallelexsitexharvestxseasonx_year

epistasisallelexsitexharvestxseasonx_year

Constructor

SimulatedEffects()

GenomicBreeding.TEBV — Type

Trial-estimate breeding values (TEBV) struct

Contains trial-estimated breeding values as generated by analyse(trials::Trials, ...).

Fields

traits: names of the traits t traits

formulae: best-fitting formula for each trait

models: best-fitting linear mixed model for each trait

df_BLUEs: best linear unbiased estimators or fixed effects table of each best fitting model

df_BLUPs: best linear unbiased predictors or random effects table of each best fitting model

df_TEBVs: breeding values of each entry per trait

Examples

julia> tebv = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], df_TEBVs=[DataFrame(x=1)]);

julia> tebv.traits
1-element Vector{String}:
 ""

GenomicBreeding.Trials — Type

Trials struct

Contains phenotype data across years, seasons, harvest, sites, populations, replications, blocks, rows, and columns

Constructor

Trials(; n::Int64 = 2, p::Int64 = 2)

Fields

phenotypes: n x t matrix of numeric phenotype data which can have missing values

traits: names of the traits t traits

years: names of the years corresponding to each row in the phenotype matrix

seasons: names of the seasons corresponding to each row in the phenotype matrix

harvests: names of the harvests corresponding to each row in the phenotype matrix

sites: names of the sites corresponding to each row in the phenotype matrix

replications: names of the replications corresponding to each row in the phenotype matrix

blocks: names of the blocks corresponding to each row in the phenotype matrix

rows: names of the rows corresponding to each row in the phenotype matrix

cols: names of the cols corresponding to each row in the phenotype matrix

entries: names of the entries corresponding to each row in the phenotype matrix

populations: names of the populations corresponding to each row in the phenotype matrix

Examples

julia> trials = Trials(n=1, t=2)
Trials(Union{Missing, Float64}[missing missing], ["", ""], [""], [""], [""], [""], [""], [""], [""], [""], [""], [""])

julia> fieldnames(Trials)
(:phenotypes, :traits, :years, :seasons, :harvests, :sites, :replications, :blocks, :rows, :cols, :entries, :populations)

Base.:== — Method

Base.:(==)(x::Genomes, y::Genomes)::Bool

Equality of Genomes structs using the hash function defined for Genomes structs.

Examples

julia> genomes_1 = genomes = Genomes(n=2,p=4);

julia> genomes_2 = genomes = Genomes(n=2,p=4);

julia> genomes_3 = genomes = Genomes(n=1,p=2);

julia> genomes_1 == genomes_2
true

julia> genomes_1 == genomes_3
false

Base.:== — Method

Base.:(==)(x::Phenomes, y::Phenomes)::Bool

Equality of Phenomes structs using the hash function defined for Phenomes structs.

Examples

julia> phenomes_1 = phenomes = Phenomes(n=2, t=4);

julia> phenomes_2 = phenomes = Phenomes(n=2, t=4);

julia> phenomes_3 = phenomes = Phenomes(n=1, t=2);

julia> phenomes_1 == phenomes_2
true

julia> phenomes_1 == phenomes_3
false

Base.:== — Method

Base.:(==)(x::SimulatedEffects, y::SimulatedEffects)::Bool

Equality of SimulatedEffects structs using the hash function defined for SimulatedEffects structs.

Examples

julia> effects_1 = SimulatedEffects();

julia> effects_2 = SimulatedEffects();

julia> effects_3 = SimulatedEffects(); effects_3.id[1] = "SOMETHING_ELSE";

julia> effects_1 == effects_2
true

julia> effects_1 == effects_3
false

Base.:== — Method

Base.:(==)(x::TEBV, y::TEBV)::Bool

Equality of TEBV structs using the hash function defined for TEBV structs.

Examples

julia> tebv_1 = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], df_TEBVs=[DataFrame(x=1)]);

julia> tebv_2 = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], df_TEBVs=[DataFrame(x=1)]);

julia> tebv_3 = TEBV(traits=["SOMETHING_ELSE"], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], df_TEBVs=[DataFrame(x=1)]);

julia> tebv_1 == tebv_2
true

julia> tebv_1 == tebv_3
false

Base.:== — Method

Base.:(==)(x::Trials, y::Trials)::Bool

Equality of Trials structs using the hash function defined for Trials structs.

Examples

julia> trials_1 = trials = Trials(n=2, t=4);

julia> trials_2 = trials = Trials(n=2, t=4);

julia> trials_3 = trials = Trials(n=1, t=2);

julia> trials_1 == trials_2
true

julia> trials_1 == trials_3
false

Base.hash — Method

Base.hash(x::Genomes, h::UInt)::UInt

Hash a Genomes struct using the entries, populations and locialleles. We deliberately excluded the allelefrequencies, and mask for efficiency.

Examples

julia> genomes = Genomes(n=2, p=2);

julia> typeof(hash(genomes))
UInt64

Base.hash — Method

Base.hash(x::Phenomes, h::UInt)::UInt

Hash a Phenomes struct.

Examples

julia> phenomes = Phenomes(n=2, t=2);

julia> typeof(hash(phenomes))
UInt64

Base.hash — Method

Base.hash(x::SimulatedEffects, h::UInt)::UInt

Hash a SimulatedEffects struct.

Examples

julia> effects = SimulatedEffects();

julia> typeof(hash(effects))
UInt64

Base.hash — Method

Base.hash(x::TEBV, h::UInt)::UInt

Hash a TEBV struct using the traits, formualae and dfTEBVs. We deliberately excluded the models, dfBLUEs, and df_BLUPs for efficiency.

Examples

julia> tebv = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], df_TEBVs=[DataFrame(x=1)]);

julia> typeof(hash(tebv))
UInt64

Base.hash — Method

Base.hash(x::Trials, h::UInt)::UInt

Hash a Trials struct.

Examples

julia> trials = Trials(n=2, t=2);

julia> typeof(hash(trials))
UInt64

Base.sum — Method

sum(effects::SimulatedEffects)::Tuple{Int64, Int64, Int64}

Sum up the simulated effects to generate the simulated phenotype values

Examples

julia> effects = SimulatedEffects();

julia> sum(effects)
1-element Vector{Float64}:
 0.0

julia> effects.additive_genetic[1] = pi;

julia> sum(effects)
1-element Vector{Float64}:
 3.141592653589793

GenomicBreeding.analyse — Method

Analyse trials

Arguments

trials: Trials struct

max_levels: maximum number of non-entry factor levels to include in the linear mixed models (default = 100)

max_time_per_model: maximum time in seconds for fitting each linear mixed model (default = 60)

verbose: Show trials analysis progress bar? (default = true)

Outputs

TEBV struct containing the trait names, the best fitting formulae, models, BLUEs, and BLUPs for each trait

Examples

julia> trials, _simulated_effects = simulatetrials(genomes = simulategenomes(n=10, verbose=false), n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=10, verbose=false);

julia> tebv = analyse(trials, max_levels=50, verbose=false);

julia> tebv.traits
3-element Vector{String}:
 "trait_1"
 "trait_2"
 "trait_3"

julia> checkdims(tebv)
true

GenomicBreeding.checkdims — Method

checkdims(genomes::Genomes)::Bool

Check dimension compatibility of the fields of the Genomes struct

Examples

julia> genomes = Genomes(n=2,p=4);

julia> checkdims(genomes)
false

julia> genomes.entries = ["entry_1", "entry_2"];

julia> genomes.loci_alleles = ["chr1\t1\tA|T\tA", "chr1\t2\tC|G\tG", "chr2\t3\tA|T\tA", "chr2\t4\tG|T\tG"];

julia> checkdims(genomes)
true

GenomicBreeding.checkdims — Method

checkdims(y::Phenomes)::Bool

Check dimension compatibility of the fields of the Phenomes struct

Examples

julia> y = Phenomes(n=2, t=2);

julia> checkdims(y)
false

julia> y.entries = ["entry_1", "entry_2"];

julia> y.traits = ["trait_1", "trait_2"];

julia> checkdims(y)
true

GenomicBreeding.checkdims — Method

checkdims(effects::SimulatedEffects)::Bool

Check dimension compatibility of the fields of the SimulatedEffects struct

Examples

julia> effects = SimulatedEffects();

julia> checkdims(effects)
true

julia> effects.id = ["beaking_change"];

julia> checkdims(effects)
false

GenomicBreeding.checkdims — Method

checkdims(y::TEBV)::Bool

Check dimension compatibility of the fields of the TEBV struct

Examples

julia> tebv = TEBV(traits=[""], formulae=[""], models=[MixedModel(@formula(y~1+(1|x)), DataFrame(y=1, x=1))], df_BLUEs=[DataFrame(x=1)], df_BLUPs=[DataFrame(x=1)], df_TEBVs=[DataFrame(x=1)]);

julia> checkdims(tebv)
true

GenomicBreeding.checkdims — Method

checkdims(trials::Trials)::Bool

Check dimension compatibility of the fields of the Trials struct

Examples

julia> trials = Trials(n=1, t=2);

julia> trials.entries = ["entry_1"];

julia> checkdims(trials)
true

julia> trials.entries = ["entering_2_entries", "instead_of_just_1"];

julia> checkdims(trials)
false

GenomicBreeding.countlevels — Method

countlevels(df::DataFrame; column_names::Vector{String})::Int64

Count the total number of factor levels across the column names provided.

GenomicBreeding.dimensions — Method

dimensions(genomes::Genomes)::Dict{String, Int64}

Count the number of entries, populations, loci-alleles combination, loci, and maximum number of alleles per locus in the Genomes struct

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> dimensions(genomes)
Dict{String, Int64} with 6 entries:
  "n_entries"      => 100
  "n_chr"          => 7
  "n_loci"         => 1000
  "n_loci_alleles" => 3000
  "n_populations"  => 1
  "max_n_alleles"  => 4

GenomicBreeding.loci — Method

loci(genomes::Genomes)::Tuple{Vector{String},Vector{Int64},Vector{Int64},Vector{Int64}}

Extract chromosome names, positions, start and end indexes of each locus across loci

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> chromsomes, positions, loci_ini_idx, loci_fin_idx = loci(genomes);

julia> length(chromsomes), length(positions), length(loci_ini_idx), length(loci_fin_idx)
(1000, 1000, 1000, 1000)

GenomicBreeding.loci_alleles — Method

loci_alleles(genomes::Genomes)::Tuple{Vector{String},Vector{Int64},Vector{String}}

Extract chromosomes, positions, and alleles across loci-allele combinations

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> chromsomes, positions, alleles = loci_alleles(genomes);

julia> length(chromsomes), length(positions), length(alleles)
(3000, 3000, 3000)

GenomicBreeding.plot — Function

plot(genomes::Genomes)::Nothing

Plot allele frequencies

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> GenomicBreeding.plot(genomes);

GenomicBreeding.readJLD2 — Method

readJLD2(type::Type, fname::String = missing)::Type

Load a core (Genomes, Phenomes, and Trials), simulation (SimulatedEffects), and model (TEBV) struct from a JLD2 file.

Examples

julia> genomes = simulategenomes(n=2, verbose=false);

julia> fname = writeJLD2(genomes);

julia> readJLD2(Genomes, fname) == genomes
true

julia> phenomes = Phenomes(n=2, t=2); phenomes.entries = ["entry_1", "entry_2"]; phenomes.traits = ["trait_1", "trait_2"];

julia> fname = writeJLD2(phenomes);

julia> readJLD2(Phenomes, fname) == phenomes
true

julia> trials = Trials(n=1, t=2); trials.entries = ["entry_1"];

julia> fname = writeJLD2(trials);

julia> readJLD2(Trials, fname) == trials
true

GenomicBreeding.readdelimited — Method

readdelimited(type::Type{Genomes}; fname::String, sep::String = "\t")::Genomes

Load a Genomes struct from a string-delimited (default=\t) file. Each row corresponds to a locus-allele combination. The first 4 columns correspond to the chromosome, position, all alleles in the locus (delimited by |), and the specific allele. The subsequency columns refer to the samples, pools, entries or genotypes.

Examples

julia> genomes = simulategenomes(n=10, verbose=false);

julia> fname = writedelimited(genomes);

julia> genomes_reloaded = readdelimited(Genomes, fname=fname);

julia> genomes == genomes_reloaded
true

GenomicBreeding.simulateeffects — Method

Simulate effects

Sample p x q effects from a multivariate normal distribution with μ~Exp(λ) and Σ=μμ'

Arguments

p: number of correlated effects to simulate (default = 2)

q: number times to simulate the correlated effects from the same distribution (default = 1)

λ: parameter of the exponential distritbution from which the means will be sampled from (default = 1.00)

seed: randomisation seed (default = 42)

Output

p x q matrix of correlated effects

Examples

julia> θ::Matrix{Float64} = simulateeffects();

julia> sum(abs.(θ - [-0.0886501800782904; -0.596478483888422])) < 0.00001
true

GenomicBreeding.simulategenomes — Method

Simulate genomes

Arguments

n: number of entries (default = 100)

l: number of loci (default = 10_000)

n_chroms: number of chromosomes (default = 7)

n_alleles: number of alleles per locus (default = 2)

max_pos: total length of the genome in base-pairs (bp) (default = 135000000)

ld_corr_50perc_kb: distance in bp at which linkage expressed as correlation between a pair of loci is at 50% (default = 1_000)

μ_β_params: the shape parameters of the Beta distribution from which the mean allele frequencies will be sampled (default = (0.5, 0.5); U-shaped distribution; you may use (2.0, 2.0) for a bell-shaped distribution)

sparsity: Proportion of missing data (default = 0.0)

seed: psuedo-random number generator seed for replicability (default = 42)

verbose: Show progress bar and plot the linkage heatmap into an svg file? (default = true)

Output

Genomes

Examples

julia> genomes = simulategenomes(n=100, l=10_000, n_alleles=3, verbose=false);

julia> length(genomes.entries)
100

julia> length(genomes.populations)
100

julia> length(genomes.loci_alleles)
20000

julia> size(genomes.allele_frequencies)
(100, 20000)

julia> mean(ismissing.(genomes.allele_frequencies))
0.0

julia> rng::TaskLocalRNG = Random.seed!(123);

julia> idx = StatsBase.sample(rng, range(1, 20_000, step=2), 250, replace = false, ordered = true);

julia> correlations = StatsBase.cor(genomes.allele_frequencies[:, idx]);

julia> correlations[10,10] == 1.00
true

julia> correlations[10,10] > correlations[10,250]
true

julia> genomes = simulategenomes(n=100, l=10_000, n_alleles=3, sparsity=0.25, verbose=false);

julia> mean(ismissing.(genomes.allele_frequencies))
0.25

GenomicBreeding.simulategenomiceffects — Method

Simulate genomic effects

Simulate additive, dominance, and epistatic effects

Arguments

genomes: Genome struct includes the n entries x p loci-alleles combinations (p = l loci x a-1 alleles)

f_additive: proportion of the l loci with non-zero additive effects on the phenotype

f_dominance: proportion of the l*f_additive additive effects loci with additional dominance effects

f_epistasis: proportion of the l*f_additive additive effects loci with additional epistasis effects

Outputs

n x 3 matrix of additive, dominance and epistasis effects per entry

p x 3 matrix of additive, dominance and epistasis effects per locus-allele combination

Examples

julia> genomes::Genomes = simulategenomes(n=100, l=2_000, n_alleles=3, verbose=false);

julia> G, B = simulategenomiceffects(genomes=genomes, f_additive=0.05, f_dominance=0.75, f_epistasis=0.25);

julia> size.([G, B])
2-element Vector{Tuple{Int64, Int64}}:
 (100, 3)
 (4000, 3)

julia> sum(B .!= 0.0, dims=1)
1×3 Matrix{Int64}:
 200  75  50

Details

The additive, dominance, and epistasis allele effects share a common exponential distribution (λ=1) from which the mean of the effects (μ) are sampled, and the covariance matrix is derived (Σ = μ * μ'; where if det(Σ)≈0 then we iteratively add 1.00 to the diagonals until it becomes invertible or 10 iterations finishes and throws an error). The non-additive or epistasis allele effects were simulated by multiplying the allele frequencies of all possible unique pairs of epistasis alleles and their effects.

GenomicBreeding.simulatetrials — Method

Simulate trials

Arguments

genomes: Genome struct includes the n entries x p loci-alleles combinations (p = l loci x a-1 alleles)

f_add_dom_epi: n_traits x 3 numeric matrix of loci proportion with additive, dominance and epistasis effects, i.e. each column refers to:

f_additive: proportion of the l loci with non-zero additive effects on the phenotype

f_dominance: proportion of the l*f_additive additive effects loci with additional dominance effects

f_epistasis: proportion of the l*f_additive additive effects loci with additional epistasis effects

n_years: Number of years (default = 2)

n_seasons: Number of seasons (default = 4)

n_harvests: Number of harvests (default = 2)

n_sites: Number of sites (default = 4)

n_replications: Number of replications (default = 2)

n_blocks: Number of blocks across the entire field layout (default = missing)

n_rows: Number of rows across the entire field layout (default = missing)

n_cols: Number of columns across the entire field layout (default = missing)

proportion_of_variance: 9 x n_traits numeric matrix of scaled/non-scaled proportion of variances allocated to genetic and environmental effects (default = missing; values will be sampled from a uniform distribution followed by a biased sample on the first row, i.e. additive effects row). The rows correspond to the variance allocated to:

additive genetic effects

dominance genetic effects

epistasis genetic effects

years effects

seasons effects

sites effects

environmental interactions

spatial interactions

GxE interactiions

seed: Randomisation seed (default = 42)

sparsity: Proportion of missing data (default = 0.0)

verbose: Show trials simulation progress bar? (default = true)

Outputs

Trials struct of simulated phenotype data

Vector of SimulatedEffects each corresponding to each trait-year-season-harvest-site-replication combination

Examples

julia> genomes::Genomes = simulategenomes(n=100, l=2_000, n_alleles=3, verbose=false);

julia> trials::Trials, vector_of_effects::Array{SimulatedEffects,1} = simulatetrials(genomes=genomes, sparsity=0.25, verbose=false);

julia> size(trials.phenotypes)
(12800, 3)

julia> size(trials.traits)
(3,)

julia> unique(trials.entries) == genomes.entries
true

julia> unique(trials.populations) == unique(genomes.populations)
true

julia> abs(mean(ismissing.(trials.phenotypes)) - 0.25) < 0.00001
true

GenomicBreeding.slice — Method

slice(genomes::Genomes;idx_entries::Vector{Int64},idx_loci_alleles::Vector{Int64})::Genomes

Count the number of entries, populations, loci, and maximum number of alleles per locus in the Genomes struct

Examples

julia> genomes = simulategenomes(n=100, l=1_000, n_alleles=4, verbose=false);

julia> sliced_genomes = slice(genomes, idx_entries=collect(1:10); idx_loci_alleles=collect(1:300));

julia> dimensions(sliced_genomes)
Dict{String, Int64} with 6 entries:
  "n_entries"      => 10
  "n_chr"          => 1
  "n_loci"         => 100
  "n_loci_alleles" => 300
  "n_populations"  => 1
  "max_n_alleles"  => 4

GenomicBreeding.trialsmodelsfomulae! — Method

trialsmodelsfomulae!(df::DataFrame; trait::String, max_levels::Int64 = 100)::Vector{String}

Define formulae for the mixed models to fit on the tabularised Trials struct. - appends interaction effects intto df - returns: + a vector of formulae as strings + a vector of the total number of non-entry factor levels

Examples

julia> trials, _simulated_effects = simulatetrials(genomes = simulategenomes(verbose=false), verbose=false);

julia> df = tabularise(trials);

julia> size(df)
(12800, 14)

julia> formulae, n_levels = trialsmodelsfomulae!(df, trait="trait_1");

julia> size(df)
(12800, 134)

julia> length(formulae)
76

julia> sum(n_levels .== sort(n_levels))
76

GenomicBreeding.writeJLD2 — Method

writeJLD2(A::Union{Genomes,Phenomes,Trials,SimulatedEffects}; fname::Union{Missing,String} = missing)::String

Save core (Genomes, Phenomes, and Trials), simulation (SimulatedEffects), and model (TEBV) structs as JLD2, a heirarchical data format version 5 (HDF5) - compatible format. Note that the extension name should be '.jld2'.

Examples

julia> genomes = simulategenomes(n=2, verbose=false);

julia> writeJLD2(genomes, fname="test_genomes.jld2")
"test_genomes.jld2"

julia> load("test_genomes.jld2")["Genomes"] == genomes
true

julia> phenomes = Phenomes(n=2, t=2); phenomes.entries = ["entry_1", "entry_2"]; phenomes.traits = ["trait_1", "trait_2"];

julia> writeJLD2(phenomes, fname="test_phenomes.jld2")
"test_phenomes.jld2"

julia> load("test_phenomes.jld2")["Phenomes"] == phenomes
true

julia> trials = Trials(n=1, t=2); trials.entries = ["entry_1"];

julia> writeJLD2(trials, fname="test_trials.jld2")
"test_trials.jld2"

julia> load("test_trials.jld2")["Trials"] == trials
true

julia> simulated_effects = SimulatedEffects();

julia> writeJLD2(simulated_effects, fname="test_simulated_effects.jld2")
"test_simulated_effects.jld2"

julia> load("test_simulated_effects.jld2")["SimulatedEffects"] == simulated_effects
true

julia> trials, _simulated_effects = simulatetrials(genomes = simulategenomes(n=10, verbose=false), n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=10, verbose=false);

julia> tebv = analyse(trials, max_levels=50, verbose=false);

julia> writeJLD2(tebv, fname="test_tebv.jld2")
"test_tebv.jld2"

julia> load("test_tebv.jld2")["TEBV"] == tebv
true

GenomicBreeding.writedelimited — Method

writedelimited(genomes::Genomes, sep::String = "	", fname::Union{Missing,String} = missing)::String

Save Genomes struct as a string-delimited (default=) file. Each row corresponds to a locus-allele combination. The first 4 columns correspond to the chromosome, position, all alleles in the locus (delimited by |), and the specific allele. The subsequency columns refer to the samples, pools, entries or genotypes.

Notes:

Extension name should be '.tsv', '.csv', or '.txt'.

Header lines and comments are prefixed by '#'.

There are 2 header lines prefixed by '#', e.g.:

header line 1: "chrom,pos,allalleles,allele,entry1,entry_2"

header line 2: "chrom,pos,allalleles,allele,population1,population_1"

Examples

julia> genomes = simulategenomes(n=2, verbose=false);

julia> writedelimited(genomes, fname="test_genomes.tsv")
"test_genomes.tsv"

GenomicBreeding.writedelimited — Method

writedelimited(phenomes::Phenomes, sep::String = "	", fname::Union{Missing,String} = missing)::String

Save Phenomes struct as a string-delimited (default=) file. Each row corresponds to a samples, pools, entries or genotypes. The first 2 columns correspond to the entry and population names. The subsequency columns refer to the traits containing the phenotype values of each entry. Note that the extension name should be '.tsv', '.csv', or '.txt'.

Notes:

Extension name should be '.tsv', '.csv', or '.txt'.

Header line and comments are prefixed by '#'.

There is 1 header line prefixed by '#', e.g.: "entry,population,trait1,trait2,trait_3"

Examples

julia> phenomes = Phenomes(n=2, t=2); phenomes.entries = ["entry_1", "entry_2"]; phenomes.traits = ["trait_1", "trait_2"];

julia> writedelimited(phenomes, fname="test_phenomes.tsv")
"test_phenomes.tsv"

GenomicBreeding.writedelimited — Method

writedelimited(trials::Trials, sep::String = "	", fname::Union{Missing,String} = missing)::String

Save Trials struct as a string-delimited (default=) file. Each row corresponds to a samples, pools, entries or genotypes. The first 10 columns correspond to:

years

seasons

harvests

sites

entries

populations

replications

blocks

rows

cols

The subsequency columns refer to the traits containing the phenotype values.

Notes:

Extension name should be '.tsv', '.csv', or '.txt'.

Header line and comments are prefixed by '#'.

There is 1 header line prefixed by '#', e.g.: "years,seasons,harvests, ..., trait1,tratit2,trait_3"

Examples

julia> trials = Trials(n=1, t=2); trials.years = ["year_1"]; trials.seasons = ["season_1"]; trials.harvests = ["harvest_1"]; trials.sites = ["site_1"]; trials.entries = ["entry_1"]; trials.populations = ["population_1"]; trials.replications = ["replication_1"]; trials.blocks = ["block_1"]; trials.rows = ["row_1"]; trials.cols = ["col_1"]; trials.traits = ["trait_1", "trait_2"];

julia> writedelimited(trials, fname="test_trials.tsv")
"test_trials.tsv"

GenomicBreeding.@string2formula — Macro

string2formula(x)

Macro to Meta.parse a string into a formula.