GenomicBreeding

Documentation for GenomicBreeding.

GenomicBreeding.GBInput

Base.:==

Base.hash

GBCore.checkdims

GenomicBreeding.assess

GenomicBreeding.clone

GenomicBreeding.extracteffects

GenomicBreeding.load

GenomicBreeding.plot

GenomicBreeding.preparegwasinputs

GenomicBreeding.prepareinputs

GenomicBreeding.submitslurmarrayjobs

GenomicBreeding.GBInput — Type

mutable struct GBInput
    fname_geno::String
    fname_pheno::String
    bulk_cv::Bool
    populations::Union{Nothing,Vector{String}}
    traits::Union{Nothing,Vector{String}}
    models::Any
    n_folds::Int64
    n_replications::Int64
    keep_all::Bool
    gwas_models::Any
    maf::Float64
    mtv::Float64
    n_iter::Int64
    n_burnin::Int64
    fname_out_prefix::String
    SLURM_job_name::String
    SLURM_account_name::String
    SLURM_partition_name::String
    SLURM_nodes_per_array_job::Int64
    SLURM_tasks_per_node::Int64
    SLURM_cpus_per_task::Int64
    SLURM_mem_G::Int64
    SLURM_time_limit_dd_hhmmss::String
    SLURM_max_array_jobs_running::Int64
    SLURM_module_load_R_version_name::String
    SLURM_module_load_BLAS_version_name::String
    verbose::Bool
end

Input struct (belongs to GBCore.AbstractGB type)

fname_geno: genotype file (see file format guide: TODO: {URL HERE})

fname_pheno: phenotype file (see file format guide: TODO: {URL HERE})

bulk_cv: perform cross-validation across all populations, i.e. disregard population grouping (Default = false)

populations: include only these populations (Default = nothing which means include all populations)

traits: include only these traits (Default = nothing which means include all traits)

models: genomic prediction model functions (Default = [ridge, bayesa]; see models list: TODO: {URL HERE})

n_folds: number of k partitions for k-fold cross-validation (Default = 5)

n_replications: number of replications for repeated k-fold cross-validation (Default = 5)

gwas_models: models to use if performning genome-wide association study (Default = [gwasols, gwaslmm])

keep_all: keep all entries upon merging genomes and phenomes potentially resulting in sparsities in both structs? (Default = false)

maf: minimum allele frequency (Default = 0.05)

mtv: minimum trait variance (Default = 1e-7)

n_iter: number of Bayesian model fitting MCMC/HMC iteration (Default = 1_500)

n_burnin: number of initial Bayesian model fitting MCMC/HMC iterations to be excluded from the posterior distribution (Default = 500)

fname_out_prefix: prefix of the output files which may include directory names (Default = "" which translates to ./GBOuput/output-<yyyymmddHHMMSS>-<3_digit_random_number>-)

SLURM_job_name: name of the Slurm job array (Default = "")

SLURM_account_name: Slurm account name (Default = "")

SLURM_partition_name: Slurm paritition to use (Default = "")

SLURM_nodes_per_array_job: number of nodes per array job (Default = 1)

SLURM_tasks_per_node: number of tasks per node (Default = 1)

SLURM_cpus_per_task: number of CPU cores per task (Default = 1)

SLURM_mem_G: maximum memroy requested in gigabytes (Default = 1)

SLURM_time_limit_dd_hhmmss: maximum computation time requested in the follwowing format: "dd-hh:mm:ss" (Default = "00-01:00:00")

SLURM_max_array_jobs_running: maximum number of array jobs which can run simultaneously (Default = 20)

SLURM_module_load_R_version_name: name of the R statistical language module and version to be used to call R::BGLR (Default = "R")

SLURM_module_load_BLAS_version_name: name of the BLAS module and version to be used to use with R::BGLR. This is non-critical as there is usually a system default to fallback to. (Default = "")

verbose: show messages (Default = true)

Base.:== — Method

Base.:(==)(x::GBInput, y::GBInput)::Bool

Equality of GBInput structs using the hash function defined for GBInput structs.

Examples

julia> input_1 = input = GBInput(fname_geno="geno1.jld2", fname_pheno="pheno1.jld2", fname_out_prefix="test1-", SLURM_job_name="slurmjob1");

julia> input_2 = input = GBInput(fname_geno="geno1.jld2", fname_pheno="pheno1.jld2", fname_out_prefix="test1-", SLURM_job_name="slurmjob1");

julia> input_3 = input = GBInput(fname_geno="geno2.jld2", fname_pheno="pheno2.jld2");

julia> input_1 == input_2
true

julia> input_1 == input_3
false

Base.hash — Method

Base.hash(x::GBInput, h::UInt)::UInt

Hash a GBInput struct using all its fields. We deliberately excluded the allele_frequencies, and mask for efficiency.

Examples

julia> input = GBInput(fname_geno="", fname_pheno="");

julia> typeof(hash(input))
UInt64

GBCore.checkdims — Method

checkdims(input::GBInput)::Bool

Check dimension compatibility of the fields of the GBInput struct

Examples

julia> input = GBInput(fname_geno="geno1.jld2", fname_pheno="pheno1.jld2");

julia> checkdims(input)
true

julia> input.models = nothing

julia> checkdims(input)
false

GenomicBreeding.assess — Method

assess(input::GBInput)::Tuple{Vector{String},Vector{String}}

Assess genomic prediction accuracy via replicated k-fold cross-validation. Outputs are saved as JLD2 (each containing a CV struct per fold, replication, and trait) and possibly text file/s containing notes describing why some jobs failed.

Example

julia> genomes = GBCore.simulategenomes(n=300, l=1_000, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> proportion_of_variance = fill(0.0, 9, 3); proportion_of_variance[1, :] .= 1.00; # 100% variance on the additive genetic effects

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, proportion_of_variance=proportion_of_variance, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;

julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, populations=["pop_1", "pop_3"], traits=["trait_1"], n_replications=2, n_folds=3, verbose=false);

julia> fnames_cvs, fnames_notes = assess(input);

julia> length(fnames_cvs) == 32, length(fnames_notes) == 0
(true, true)

GenomicBreeding.clone — Method

clone(x::GBInput)::GBInput

Clone a GBInput object

Example

julia> input = GBInput(fname_geno="geno1.jld2", fname_pheno="pheno1.jld2");

julia> copy_input = clone(input);

julia> input == copy_input
true

GenomicBreeding.extracteffects — Method

extracteffects(input::GBInput)::Vector{String}

Extract allele effects by fitting the models without cross-validation. The output are filenames of tab-delimited files containing the allele effects per model, trait and population.

Example

julia> genomes = GBCore.simulategenomes(n=300, l=1_000, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> proportion_of_variance = fill(0.0, 9, 3); proportion_of_variance[1, :] .= 1.00; # 100% variance on the additive genetic effects

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, proportion_of_variance=proportion_of_variance, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;

julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, populations=["pop_1", "pop_3"], traits=["trait_1"], n_replications=2, n_folds=3, verbose=false);

julia> fname_effects = extracteffects(input);

julia> length(fname_effects) == 6
true

GenomicBreeding.load — Method

load(input::GBInput)::Tuple{Genomes, Phenomes}

Load, merge and filter genotype and phenotype data

Example

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;
    
julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, populations=["pop_1", "pop_3"], traits=["trait_1"], verbose=false);

julia> genomes, phenomes = load(input);

julia> length(unique(genomes.populations)) == length(unique(phenomes.populations)) == 2
true

julia> length(phenomes.traits) == 1
true

julia> rm.([fname_geno, fname_pheno]);

GenomicBreeding.plot — Method

plot(;
    input::GBInput,
    skip_genomes::Bool = false,
    skip_phenomes::Bool = false,
    format::String = "svg",
    plot_size::Tuple{Int64,Int64} = (600, 450),
    overwrite::Bool = false,
)::String

Plot genomes, phenomes, and CVs, if present.

Example

julia> genomes = GBCore.simulategenomes(n=300, l=1_000, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;

julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;
    
julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, SLURM_cpus_per_task=6, SLURM_mem_G=5, fname_out_prefix="GBOutput/test-", verbose=false);

julia> GenomicBreeding.plot(input=input, format="png", plot_size = (700, 525))
"GBOutput/plots"

julia> GenomicBreeding.plot(input=input, format="png", plot_size = (700, 525), overwrite=true, skip_genomes=true)
"GBOutput/plots"

GenomicBreeding.preparegwasinputs — Method

preparegwasinputs(input::GBInput)::Vector{GBInput}

Prepare GBInputs for Slurm array jobs

Example

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;
    
julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, verbose=false);

julia> inputs = preparegwasinputs(input);

julia> length(inputs) == 30
true

julia> rm.([fname_geno, fname_pheno]);

GenomicBreeding.prepareinputs — Method

prepareinputs(input::GBInput)::Vector{GBInput}

Prepare GBInputs for Slurm array jobs

Example

julia> genomes = GBCore.simulategenomes(n=300, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;
    
julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, verbose=false);

julia> inputs = prepareinputs(input);

julia> length(inputs) == 30
true

julia> rm.([fname_geno, fname_pheno]);

GenomicBreeding.submitslurmarrayjobs — Method

submitslurmarrayjobs(; input::GBInput, analysis::Function)::String

Assess genomic prediction accuracy via replicated k-fold cross-validation. Outputs are saved as JLD2 (each containing a CV struct per fold, replication, and trait) and possibly text file/s containing notes describing why some jobs failed. Note that you will be prompted to enter YES to proceed with Slurm job submission after showing you the job details to review and confirm.

Example

<!– ```jldoctest; setup = :(using GBCore, GBIO, GenomicBreeding, StatsBase, DataFrames) –>

julia> genomes = GBCore.simulategenomes(n=300, l=1_000, verbose=false); genomes.populations = StatsBase.sample(string.("pop_", 1:3), length(genomes.entries), replace=true);

julia> trials, _ = GBCore.simulatetrials(genomes=genomes, n_years=1, n_seasons=1, n_harvests=1, n_sites=1, n_replications=1, verbose=false);

julia> phenomes = extractphenomes(trials);

julia> fname_geno = try writedelimited(genomes, fname="test-geno.tsv"); catch; rm("test-geno.tsv"); writedelimited(genomes, fname="test-geno.tsv"); end;

julia> fname_pheno = try writedelimited(phenomes, fname="test-pheno.tsv"); catch; rm("test-pheno.tsv"); writedelimited(phenomes, fname="test-pheno.tsv"); end;

julia> input = GBInput(fname_geno=fname_geno, fname_pheno=fname_pheno, SLURM_cpus_per_task=6, SLURM_mem_G=5, fname_out_prefix="GBOutput/test-", verbose=false);

julia> outdir = submitslurmarrayjobs(input=input, analysis=assess)
GBOutput

julia> run(`squeue`)